Move ProcessType (Processable) to its own module
